#izmene bagova

+++ IzdajMe/src/main/java/com/izdajMe/izdajMe/services/UserServiceImpl.java
@@ -40,6 +40,8 @@ public class UserServiceImpl implements UserService {
     private ShipReservationService shipReservationService;
     @Autowired
     private FavorReservationService favorReservationService;
+    @Autowired
+    private ConcurentWatcherRepository concurentWatcherRepository;
 
     @Override
     public List<User> getAllUsers() {
@@ -262,10 +264,21 @@ public class UserServiceImpl implements UserService {
     }
 
     public Boolean declineUser(String text) {
-        User user = userRepository.findById((long) Integer.parseInt(text.split(" ")[0])).get();
-        userRepository.deleteById((long) Integer.parseInt(text.split(" ")[0]));
-        sendNotificationFromAdminForFailedRegistration(user, text.split(" ")[1]);
-        return true;
+        if(!concurentWatcherRepository.findByTableName("AnswerToRegistrationRequest").getWriting()) {
+            ConcurentWatcher cw = concurentWatcherRepository.findByTableName("AnswerToRegistrationRequest");
+            cw.setWriting(true);
+            concurentWatcherRepository.save(cw);
+
+            User user = userRepository.findById((long) Integer.parseInt(text.split(" ")[0])).get();
+            userRepository.deleteById((long) Integer.parseInt(text.split(" ")[0]));
+            sendNotificationFromAdminForFailedRegistration(user, text.split(" ")[1]);
+
+            cw.setWriting(false);
+            concurentWatcherRepository.save(cw);
+
+            return true;
+        } else
+            return false;
     }
 
     public Boolean deleteUser(long id) {
@@ -603,11 +616,22 @@ public class UserServiceImpl implements UserService {
     }
 
     public Boolean acceptUser(long id){
-        User user = userRepository.findById(id).get();
-        user.setVerified(true);
-        userRepository.save(user);
-        sendNotificationFromAdminForSuccessRegistration(user);
-        return true;
+        if(!concurentWatcherRepository.findByTableName("AnswerToRegistrationRequest").getWriting()) {
+            ConcurentWatcher cw = concurentWatcherRepository.findByTableName("AnswerToRegistrationRequest");
+            cw.setWriting(true);
+            concurentWatcherRepository.save(cw);
+
+            User user = userRepository.findById(id).get();
+            user.setVerified(true);
+            userRepository.save(user);
+            sendNotificationFromAdminForSuccessRegistration(user);
+
+            cw.setWriting(false);
+            concurentWatcherRepository.save(cw);
+
+            return true;
+        } else
+            return false;
     }
 
     private void sendNotificationFromAdminForSuccessRegistration(User user) throws MailException{
